+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 -----------------------
 
 * Add new method :meth:`pymaf.MafFrame.plot_waterfallp`.
+* Add new method :meth:`pymaf.MafFrame.plot_patient_tmb`.
 
 0.19.0 (2021-07-31)
 -------------------